requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Flask==3.0.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==10.4.0
pixoo==0.9.2
PyYAML==6.0.2
requests==2.32.5
urllib3==2.5.0
Werkzeug==3.1.3
